Console: Move interpreter.py to utils
