+++ spyder/plugins/console/utils/interpreter.py
